+++ src/pipecat/processors/aggregators/gated.py
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: BSD 2-Clause License
 #
 
-from typing import List
+from typing import List, Tuple
 
 from pipecat.frames.frames import Frame, SystemFrame
 from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
@@ -40,18 +40,25 @@ class GatedAggregator(FrameProcessor):
     Goodbye.
     """
 
-    def __init__(self, gate_open_fn, gate_close_fn, start_open):
+    def __init__(self, gate_open_fn, gate_close_fn, start_open,
+                 direction: FrameDirection = FrameDirection.DOWNSTREAM):
         super().__init__()
         self._gate_open_fn = gate_open_fn
         self._gate_close_fn = gate_close_fn
         self._gate_open = start_open
-        self._accumulator: List[Frame] = []
+        self._direction = direction
+        self._accumulator: List[Tuple[Frame, FrameDirection]] = []
 
     async def process_frame(self, frame: Frame, direction: FrameDirection):
         await super().process_frame(frame, direction)
 
         # We must not block system frames.
         if isinstance(frame, SystemFrame):
+            await self.push_frame(frame, direction)
+            return
+
+        # Ignore frames that are not following the direction of this gate.
+        if direction != self._direction:
             await self.push_frame(frame, direction)
             return
 
@@ -67,8 +74,8 @@ class GatedAggregator(FrameProcessor):
 
         if self._gate_open:
             await self.push_frame(frame, direction)
-            for frame in self._accumulator:
-                await self.push_frame(frame, direction)
+            for (f, d) in self._accumulator:
+                await self.push_frame(f, d)
             self._accumulator = []
         else:
-            self._accumulator.append(frame)
+            self._accumulator.append((frame, direction))